+++ odb/test.py
@@ -1,7 +1,7 @@
 import unittest
 import time
 
-from odb import ODB
+from odb import ODB, TagExists
 
 
 class TestCommit(unittest.TestCase):
@@ -13,7 +13,7 @@ class TestCommit(unittest.TestCase):
         ODB(db).createdb()
 
     def test_simple_commit(self):
-        """ first simple scenario with snapshot and revert
+        """ first simple scenario with commit and revert
         """
         odb = ODB(self.db)
         # init
@@ -22,15 +22,15 @@ class TestCommit(unittest.TestCase):
         self.assertEqual(odb.parent(), 0)
         # init
         self.assertEqual(odb.init(), 1)
-        # snapshot
-        odb.snapshot()
+        # commit
+        odb.commit()
         self.assertEqual(odb.revision(), 2)
         self.assertEqual(odb.parent(), 1)
         # the db name remains the same
         self.assertEqual(self.db, odb.db)
 
-        # snapshot
-        odb.snapshot()
+        # commit
+        odb.commit()
         self.assertEqual(odb.revision(), 3)
         self.assertEqual(odb.parent(), 2)
 
@@ -39,8 +39,8 @@ class TestCommit(unittest.TestCase):
         self.assertEqual(odb.revision(), 3)
         self.assertEqual(odb.parent(), 2)
 
-        # snapshot
-        odb.snapshot()
+        # commit
+        odb.commit()
         self.assertEqual(odb.revision(), 4)
         self.assertEqual(odb.parent(), 3)
         # just check that init again doesn't hurt
@@ -52,20 +52,45 @@ class TestCommit(unittest.TestCase):
         self.assertEqual(odb.parent(), 2)
         self.assertEqual(self.db, odb.db)
 
-        # snapshot
-        odb.snapshot()
+        # commit
+        odb.commit()
         self.assertEqual(odb.revision(), 5)
         self.assertEqual(odb.parent(), 4)
 
         # list all revisions
         revs = odb.log()
-        self.assertEqual(revs[-1]['revision'], '1')
+        self.assertEqual(revs[-1]['revision'], 1)
         self.assertEqual(revs[1]['db'], self.db + '*4')
-        self.assertEqual(revs[0]['revision'], '5')
+        self.assertEqual(revs[0]['revision'], 5)
+
+        # tag revision 3 and 5
+        odb.tag('v1', 3)
+        odb.tag('v2')
+        self.assertRaises(TagExists, odb.tag, 'v1', 4)
+
+        # tags appear in the log
+        self.assertEqual(odb.log()[2]['tag'], 'v1')
+        self.assertEqual(odb.log()[0]['tag'], 'v2')
+
+        # delete tag
+        odb.tag('v2', delete=True)
+        self.assertEqual(odb.log()[0].get('tag'), None)
+
+        # we can revert to a tag
+        odb.revert(tag='v1')
+        revs = odb.log()
+        self.assertEqual(revs[0]['revision'], 5)
+        self.assertEqual(revs[0]['parent'], 3)
+        # the current db has not the same tag
+        self.assertEqual(odb.get('tag'), None)
+        # even after commit, no tag
+        odb.tag('v3')
+        odb.commit()
+        self.assertEqual(odb.get('tag'), None)
 
         # purge without confirmation does nothing
         odb.purge('all')
-        self.assertEqual(len(odb.log()), 5)
+        self.assertEqual(len(odb.log()), 6)
 
         # check for real
         odb.purge('all', confirm=True)
@@ -74,6 +99,7 @@ class TestCommit(unittest.TestCase):
     def tearDown(self):
         """ cleanup
         """
-        # clean the current db
+        # drop everything (even if tests failed)
         odb = ODB(self.db)
+        odb.purge('all', confirm=True)
         odb.dropdb()